game: building interface of SyncableItem and syncableMaxU8/U16 implementations to generate ASM routines for writes to the SNES

+++ games/alttp/player.go
@@ -10,7 +10,7 @@ type Player struct {
 	g *Game
 
 	Index int
-	TTL   uint8
+	TTL   int
 
 	Team uint8
 	Name string
